# apps/platform-api/specs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Blog API Documentation",
    "version": "1.0.0",
    "description": "API Documentation"
  },
  "paths": {
...
    "/api/v1/role": {
      "get": {
        "summary": "列出所有角色",
        "tags": [
          "Roles"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": true,
            "schema": {
              "type": "object"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": true,
            "schema": {
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "角色列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "description": {
                            "type": "object"
                          },
                          "created_at": {
                            "type": "string"
                          },
                          "updated_at": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "description",
                          "created_at",
                          "updated_at"
                        ]
                      }
                    },
                    "total": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "data",
                    "total"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "创建新角色",
        "tags": [
          "Roles"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "description": {
                    "type": "string"
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "角色创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "affectedRows": {
                      "type": "number"
                    },
                    "id": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "affectedRows"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/role/{id}": {
      "get": {
        "summary": "根据ID获取角色",
        "tags": [
          "Roles"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "角色详情",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "description": {
                      "type": "object"
                    },
                    "created_at": {
                      "type": "string"
                    },
                    "updated_at": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "description",
                    "created_at",
                    "updated_at"
                  ]
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "更新角色",
        "tags": [
          "Roles"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "角色更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "affectedRows": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "affectedRows"
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "删除角色",
        "tags": [
          "Roles"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "角色删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "affectedRows": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "affectedRows"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/user": {
      "get": {
        "summary": "列出所有用户（管理员）",
        "tags": [
          "Admin",
          "User"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": true,
            "schema": {
              "type": "object"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": true,
            "schema": {
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "用户列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "username": {
                            "type": "string"
                          },
                          "email": {
                            "type": "object"
                          },
                          "loginType": {
                            "type": "string",
                            "enum": [
                              "password",
                              "google",
                              "github",
                              "wechat",
                              "email"
                            ]
                          },
                          "preferredLocale": {
                            "type": "string",
                            "enum": [
                              "zh-CN",
                              "en"
                            ]
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "NORMAL",
                              "MUTED",
                              "BANNED",
                              "DELETED"
                            ]
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "updatedAt": {
                            "type": "string"
                          },
                          "lastLoginAt": {
                            "type": "object"
                          }
                        },
                        "required": [
                          "id",
                          "username",
                          "email",
                          "loginType",
                          "preferredLocale",
                          "status",
                          "createdAt",
                          "updatedAt",
                          "lastLoginAt"
                        ]
                      }
                    },
                    "total": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "data",
                    "total"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}